Update tests for better coverage
